requirements.txt
flask==3.0.3
flask-cors==4.0.1
google-generativeai==0.8.3
gunicorn==22.0.0